pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "excelmanus"
version = "1.6.2"
description = "基于大语言模型的 Excel 智能代理框架，支持 CLI / API 双模式运行"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "numpy>=1.24.0",
    "httpx>=0.24.0",
    "cryptography>=41.0.0",
    "mcp>=1.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
cli = [
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "hypothesis>=6.0.0",
]
all = [
    "excelmanus[cli]",
    "excelmanus[web]",
]

[project.scripts]
excelmanus = "excelmanus.cli:main"
excelmanus-api = "excelmanus.api:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["excelmanus*"]
